working waypoint node
