+++ src/waypoints/package.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<package format="3">
+  <name>waypoints</name>
+  <version>0.1.0</version>
+  <description>Waypoint routing for autonomous bike navigation.</description>
+  <maintainer email="dcl252@cornell.edu">Dalton Luce</maintainer>
+  <license>MIT</license>
+
+  <buildtool_depend>ament_python</buildtool_depend>
+
+  <exec_depend>rclpy</exec_depend>
+  <exec_depend>std_msgs</exec_depend>
+
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
+</package>
